feat: use local versioning.
Getting the version from git is nice, but we don't use one git repo/document, so
bumping the *whole* version no. every time we change one document is silly.
Stored like this we can rebuild only when it changes in ci (pulling the other
documents in from the last release).